requirements.txt
# Core Agent Framework
google-adk>=1.20.0
ag-ui-adk>=0.4.1  # Core middleware for AG-UI
google-genai>=0.3.0
python-dotenv>=1.0.0

# Data & Technical Analysis
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.0
PyWavelets>=1.4.0  # Required for signal_processing
scipy>=1.10.0      # Required for signal.detrend/filter
requests>=2.31.0

# Frontend & API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
copilotkit>=0.1.0
python-multipart